+++ src/lib/BookingSelector.js
@@ -991,7 +991,7 @@ export default class BookingSelector extends React.Component<PropsType, StateTyp
 
   renderDateCell = (time: Date, selected: boolean, blocked: boolean): React.Node => {
     if (this.props.renderDateCell) {
-      return this.props.renderDateCell(time, selected, blocked)
+      return this.props.renderDateCell(new Date(time.getTime()), selected, blocked)
     }
 
     return (